start restructuring containers
