+++ src/components/LoginForm/index.test.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { StaticRouter } from "react-router";
 import { Provider } from "react-redux";
 import { configureStore } from "util/configureStore";
-import LoginContainer from "./";
+import LoginContainer from "../../index";
 
 describe("<LoginContainer />", () => {
   it("renders without crashing", () => {
