+++ src/components/SignupForm/index.js
@@ -1,1 +1,13 @@
-export { default } from "./SignupForm";
+import { connect } from "react-redux";
+import SignupForm from "./SignupForm";
+import { doSignUp } from "../../thunks/signup";
+
+const mapDispatchToProps = (dispatch, { history }) => ({
+  callAPI(values) {
+    return dispatch(doSignUp(values.email, values.password));
+  }
+});
+
+const Signup = connect(null, mapDispatchToProps)(SignupForm);
+
+export default Signup;
