+++ src/components/Transactions/index.js
@@ -1,1 +1,4 @@
-export { default } from "./Transactions";
+import Transactions from "./Transactions";
+import { withTransactions } from "../../containers/transactions";
+
+export default withTransactions(Transactions);
