--- src/containers/Signup/Signup.js
@@ -1,14 +1,0 @@
-import { connect } from "react-redux";
-import { withRouter } from "react-router-dom";
-import SignupForm from "components/SignupForm";
-import { doSignUp } from "thunks/signup";
-
-const mapDispatchToProps = (dispatch, { history }) => ({
-  callAPI(values) {
-    return dispatch(doSignUp(values.email, values.password));
-  }
-});
-
-const Signup = connect(null, mapDispatchToProps)(SignupForm);
-
-export default withRouter(Signup);
