--- src/containers/Users/Users.js
@@ -1,6 +1,0 @@
-import { connect } from "react-redux";
-import Users from "components/Users";
-
-const Container = connect()(Users);
-
-export default Container;
