--- src/containers/Users/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Users";
