+++ src/containers/eos-account/index.js
