+++ src/containers/eos-account/with-eos-account-name.js
