+++ src/containers/transactions/index.js
@@ -1,6 +1,4 @@
 import { connect } from "react-redux";
-import { withRouter } from "react-router-dom";
-import Transactions from "components/Transactions";
 import { selectRecentTransactions } from "../../redux-modules/transactions/transactions-selectors";
 
 // TODO parameterize props
@@ -8,6 +6,6 @@ const mapStateToProps = state => ({
   data: selectRecentTransactions(state)
 });
 
-const Container = connect(mapStateToProps)(Transactions);
+export const withTransactions = connect(mapStateToProps);
 
-export default withRouter(Container);
+export default withTransactions;
