+++ src/routes/Accounts/Accounts.js
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Helmet } from "react-helmet";
 import ConnectedEOSAccounts from "components/AddEOSAccountForm";
 import EOSAccount from "../../components/EOSAccount";
-import Notifications from "../../containers/Notifications/index";
+import { AppNotifications as Notifications } from "../../components/Notification";
 import { selectEOSAccountName } from "../../redux-modules/eos-account/account-selectors";
 import { connect } from "react-redux";
