+++ src/routes/Signup/Signup.js
@@ -1,19 +1,13 @@
 import * as React from "react";
-import Notifications from "containers/Notifications";
-import Container from "containers/Signup";
+import { AppNotifications as Notifications } from "../../components/Notification";
+import Container from "../../components/SignupForm";
 
-const Signup = ({
-  children,
-  handleClose,
-}) => (
+const Signup = ({ children, handleClose }) => (
   <div className="login">
     <div className="login-header modal-header">
       <h3 className="title is-3">Create your Wallet account</h3>
 
-      <button
-        className="js-modal-close"
-        onClick={handleClose}
-      >
+      <button className="js-modal-close" onClick={handleClose}>
         x
       </button>
     </div>
